fix: 가이드 옵션 세션 스토리지 저장 수정
